+++ include/inputs/Raton.hpp
@@ -2,21 +2,21 @@
 #pragma once
 
 struct Raton {
-	double x, y;
+	int x, y;
 	bool leftButton, rightButton, middleButton;
 	bool capturePointer;
 	int wheel;
 	Raton();
-	Raton(double valX, double valY);
-	void getPointer(double &valX, double &valY);
-	double getPointerX();
-	double getPointerY();
-	double getCenterPointerX(int w);
-	double getCenterPointerY(int h);
+	Raton(int valX, int valY);
+	void getPointer(int &valX, int &valY);
+	int getPointerX();
+	int getPointerY();
+	int getCenterPointerX(int w);
+	int getCenterPointerY(int h);
 	bool getCapturePointer();
 	void toggleCapturePointer();
 
-	void setPointer(double valX, double valY);
+	void setPointer(int valX, int valY);
 	void getButtons(bool &leftButton, bool &rightButton, bool &middleButton);
 	void setButtons(bool leftButton, bool rightButton, bool middleButton);
 	void setWheel(int wheel, int direction, int x, int y);
